First part of selective synchronization improved

+++ peerbox/src/main/java/org/peerbox/watchservice/states/InitialState.java
@@ -97,6 +97,10 @@ public class InitialState extends AbstractActionState {
 			updateTimeAndQueue();
 			return changeStateOnLocalCreate();
 		} else {
+			if(moveSource.isSynchronized() == false){
+				updateTimeAndQueue();
+				return changeStateOnLocalCreate();
+			}
 			fileTree.deleteFile(moveSource.getPath());
 			eventManager.getFileComponentQueue().remove(file);
 			if (moveSource.isUploaded()) {
